% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindspiker-package.R
\docType{package}
\name{blindspiker-package}
\alias{blindspiker}
\alias{blindspiker-package}
\title{blindspiker: Report on bioassay laboratory blind spike sample analyses}
\description{
A blind spike program provides samples to a laboratory in order to perform quality control (QC) checks. The samples provided are of a known quantity to the tester. The laboratory is typically uninformed of that the sample provided is a QC sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markhogue/blindspiker}
  \item Report bugs at \url{https://github.com/markhogue/blindspiker/issues}
}

}
\author{
\strong{Maintainer}: Mark Hogue \email{mark.hogue.chp@gmail.com}

}
\keyword{internal}
