% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_analytes.R
\name{bs_analytes}
\alias{bs_analytes}
\title{Tabulate analytes submitted}
\usage{
bs_analytes(my_dir, df = bs_df)
}
\arguments{
\item{my_dir}{directory where data files are stored in format 'C:/my_data_loc/'.
Note the forward slashes.}

\item{compiled_data}{Data compiled with blindspiker::get_data.R. Defaults to 'df.tsv'}
}
\description{
Provide table of analyte mixtures
}
