% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labvals.R
\docType{data}
\name{labvals}
\alias{labvals}
\title{Example Laboratory Values}
\format{
## `labvals`
A data frame with 225 rows and 8 columns:
\describe{
  \item{sample_ID}{Unique sample identifier}
  \item{analyte}{what is being analyzed}
  \item{result_date}{date of analysis}
  \item{k}{uncertainty coverage factor}
  \item{result}{value of analysis}
  \item{unc}{measurement uncertainty in result units}
  \item{units}{units of measure}
  \item{det_lvl}{minimum level of detection}
  }
}
\usage{
labvals
}
\description{
Example Laboratory Values
}
\keyword{datasets}
