% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cc.R
\name{plot_cc}
\alias{plot_cc}
\title{Plot Control Chart}
\usage{
plot_cc(my_dir, df = bs_df, isotope)
}
\arguments{
\item{my_dir}{directory where data files are stored in format 'C:/my_data_loc/'.
Note the forward slashes.}

\item{isotope}{What is being analyzed}

\item{compiled_data}{Data compiled with blindspiker::get_data.R. Defaults to 'df.tsv'}
}
\description{
Plot Control Charts by Isotope
}
