% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run.R
\name{plot_run}
\alias{plot_run}
\title{Plot Run Chart}
\usage{
plot_run(
  select_analyte,
  dat = bs_df,
  version = "ratio",
  matrix = "sample",
  log = "n"
)
}
\arguments{
\item{select_analyte}{the selected analyte for this run chart}

\item{version}{The run chart is either shown with `original` units, default, or
with the result shown as a `ratio` to the spike value.}

\item{matrix}{what the analyte is suspended in. For example, 'water'. Default
is 'sample'.}

\item{log}{Set log = "y" to make the y-axis a log scale. Default is "n".}

\item{df}{data frame with all data needed as described in `get_data`.
Default is `bs_df`.}
}
\description{
`plot_run()` produces a Run Chart of the selected analyte. The analyte
is selected from the analyte set provided. Results are plotted with error
bars (uncertainty with coverage factor of 2). When original results are
plotted, the spike values are shown with a small salmon-colored "+".
}
\examples{
example_spike_data <- system.file('extdata', 'spikevals.csv', package = 'blindspiker')
example_lab_data <- system.file('extdata', 'labvals.csv', package = 'blindspiker')
example_df <- get_data(spike_data = example_spike_data, lab_data = example_lab_data)
plot_run(select_analyte = 'unknownium', df = example_df)

}
