% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikevals.R
\docType{data}
\name{spikevals}
\alias{spikevals}
\title{Example Spike Values}
\format{
## `spikevals`
A data frame with 225 rows and 8 columns:
\describe{
  \item{sample_ID}{Unique sample identifier}
  \item{analyte}{what is being analyzed}
  \item{spike_value}{measure of sample spiking}
  \item{sv_unc}{uncertainty spike value}
  \item{sv_k}{coverage factor of uncertainty}
  \item{spike_unit}{unit of measure}
  \item{units}{units of measure}
  \item{provider_lab}{origin of spike sample}
  \item{submission_date}{date provided to lab}
  }
}
\usage{
spikevals
}
\description{
Example Spike Values
}
\keyword{datasets}
