% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Collect data from three spreadsheets. User must assign name of data frame generated.
All functions default to look for data as `bs_df`}
\usage{
get_data(
  my_dir,
  lab_results,
  spikeval = "Spiked Values.xlsx",
  received_dates = "EBL Received Dates.xlsx"
)
}
\arguments{
\item{my_dir}{directory where data files are stored in format 'C:/my_data_loc/'.
Note the forward slashes.}

\item{lab_results}{name of the file containing the spike values. This file must include
the following columns in the range of A-I: SAMPLE_ID, RESULT_DATE, ISOTOPE, ACTIVITY,
ERROR, and DET_LEVEL. No default.}

\item{spikeval}{name of the file containing the spike values.
This file must include the following in columns A-E: SAMPLE_ID, Received, Result,
ISOTOPE, and SPIKED_VALUE. Default file name is 'Spiked Values.xlsx'}

\item{received_dates}{name of the file containing the lab's received and posted dates. This
file must include the following columns in the range of A-C: lngSampleNumber,
dtmDateLogged, and dtmReceiveDate. Default file name = 'EBL Received Dates.xlsx'}
}
\value{
data frame containing all needed data to be used in subsequent functions.
# add example
}
\description{
This function collects data from three different sources.
All data is read-in directly from the .xlsx files. Director and file names must be entered.
}
\details{
The following data treatment is performed:

The error is divided by 2 so that the package will work with single coverage uncertainty terms.

Unused data columns are removed. Lab data (`lab_results` and `received_dates`) are combined.
Then the data is `left_join`ed by SAMPLE_ID to `spikeval` (retaining all rows in `spikeval`).

Dates are converted to date formats. Isotopes are changed to proper capitalization.

False negative and false positive values are flagged with an added `err_type` column.
}
