% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{quantile-quantile plot}
\usage{
plot_qq(select_analyte, dat = bs_df)
}
\arguments{
\item{select_analyte}{the selected analyte for this run chart}

\item{dat}{data frame with all data needed as described in `bs_prep_and_analysis`.
Default is `bs_df`.}
}
\description{
QQ plots by isotope
}
\examples{
example_spike_data <- system.file('extdata', 'spikevals.csv', package = 'blindspiker')
example_lab_data <- system.file('extdata', 'labvals.csv', package = 'blindspiker')
example_df <- bs_prep_and_analysis(spike_data = example_spike_data, lab_data = example_lab_data)
plot_qq(select_analyte = 'unknownium', dat = example_df)

}
