% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{quantile-quantile plot}
\usage{
plot_qq(my_dir, df = bs_df, isotope)
}
\arguments{
\item{my_dir}{directory where data files are stored in format 'C:/my_data_loc/'.
Note the forward slashes.}

\item{isotope}{What is being analyzed}

\item{compiled_data}{Data compiled with blindspiker::get_data.R. Defaults to 'df.tsv'}
}
\description{
QQ plots by isotope
}
