% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tat.R
\name{plot_tat}
\alias{plot_tat}
\title{Plot turnaround time}
\usage{
plot_tat(my_dir, df = bs_df, isotope, target_days = 60)
}
\arguments{
\item{my_dir}{directory where data files are stored in format 'C:/my_data_loc/'.
Note the forward slashes.}

\item{isotope}{What is being analyzed}

\item{target_days}{The target turnaround time in days. Default = 60.}

\item{compiled_data}{Data compiled with blindspiker::get_data.R.}
}
\description{
Plot time for labratory analysis by date (result date from lab - lab received date)
}
